+++ index.html
@@ -229,5 +229,5 @@
 
 </body>
 <script src="js/functions.js"></script>
-<script src="js/main.js"></script>
+<script type="module" src="js/main.js" ></script>
 </html>
